lib: mfu Add Modem Firmware Update library
Modem Firmware Update (mfu) contains functionality for verifying
the integrity of an update image stored in (external) flash memory,
and for installing the update image into the nRF91 modem.

A modem update image is a binary file which is a compressed
representation of the official mfw_nrf9160_x.y.z.zip file releases.

A separate python script is used to generate the modem update binary.

Tinycrypt or mbedtls is used for sha256 generation.
Tinycrypt is used for ecdsa signature verification.

Signed-off-by: Audun Korneliussen <audun.korneliussen@nordicsemi.no>